Compute word scores by manufacturer. Remove close decisions.
A brand with very little representation might result in a high
entropy score for, e.g., a family name. This can result in
over-confidence when that name is seen. By computing entropy by
manufacturer, we distinguish between cases where the name is still
uncommon, and cases where half of the products (or more) have that
name.

This still leaves times where we make bad guesses (or guesses at all)
for a listing that's similar to known products, but isn't a known
product itself. In these cases, we end up with multiple products
with similar scores. If we find that the top two scores are too close
it's a strong sign that we don't actually know what the product is,
and should refrain from guessing.